feature: добавлены аргументы для вызова из командной строки для всех функций

+++ src/models/optimization/pruning.py
@@ -1,4 +1,7 @@
+import argparse
 import os
+import sys
+import traceback
 
 import torch
 import torch.nn as nn
@@ -7,10 +10,40 @@ from dotenv import load_dotenv
 from torch.nn.utils import prune
 from ultralytics import YOLO
 
+# Parse arguments
+parser = argparse.ArgumentParser(
+    description="Transform YOLO model to special format"
+)
+parser.add_argument(
+    "--model",
+    type=str,
+    default="1",
+    help="""
+    Model choice: 
+    1 for trained model, 
+    2 for optimized model, 
+    'path/to/model.pt' for custom path
+    """,
+)
+parser.add_argument(
+    "--config",
+    type=str,
+    default="yolo8_baseline.yaml",
+    help="Config model choice (default: yolo8_baseline.yaml)",
+)
+parser.add_argument(
+    "--format",
+    type=str,
+    default="torchscript",
+    help="Model format: onnx, torchscript, engine (default: torchscript)",
+)
+parse_args = parser.parse_args()
 load_dotenv()
+
 
 # Set directory
 os.chdir(os.getenv("HOME_DIR"))
+
 
 # Set model.yaml path
 # Create your yaml config file model
@@ -18,12 +51,11 @@ os.chdir(os.getenv("HOME_DIR"))
 # project_results_name: example_project
 # optimized_project_results_name: example_project_optimized
 # selected_classes: [class0, class1, class2, ..., classN]
-handle_model_yaml = "yolo8_baseline.yaml"  # handle_model.yaml
 yaml_path = os.path.join(
     os.getenv("HOME_DIR"),
     "config",
     "models",
-    handle_model_yaml,
+    parse_args.config,
 )
 with open(yaml_path, "r") as file:
     args = yaml.safe_load(file)
@@ -41,33 +73,52 @@ OUTPUT_DIR = os.path.join(
 )  # Result directory
 
 
+# Select model based on choice
+if parse_args.model == "1":
+    model_path = os.path.join(PROJECT_DIR, "train", "weights", "best.pt")
+elif parse_args.model == "2":
+    model_path = os.path.join(PROJECT_DIR, "optimized", "best_optimized.pt")
+elif parse_args.model not in ["1", "2"]:
+    model_path = parse_args.model
+else:
+    raise ValueError(
+        "Invalid model choice."
+    )
+
+if not os.path.exists(model_path):
+    raise FileNotFoundError(f"Model not found at {model_path}")
+
+
 # Load model
-model_path = os.path.join(
-    PROJECT_DIR,
-    "train",
-    "weights",
-    "best.pt",
-)
 model = YOLO(model_path, task="detect", verbose=True)
 
 
 # Pruning 0.1 weights
-for name, module in model.model.named_modules():
-    if isinstance(module, nn.Conv2d) or isinstance(module, nn.Linear):
-        prune.l1_unstructured(module, name="weight", amount=0.1)
-        prune.remove(module, "weight")
+try:
+    for name, module in model.model.named_modules():
+        if isinstance(module, nn.Conv2d) or isinstance(module, nn.Linear):
+            prune.l1_unstructured(module, name="weight", amount=parse_args.prune_amount)
+            prune.remove(module, "weight")
 
 
-# Saving model
-os.makedirs(OUTPUT_DIR, exist_ok=True)
-ckpt = {
-    "model": model.model,
-    "train_args": {},
-}
-torch.save(
-    ckpt,
-    os.path.join(
-        OUTPUT_DIR,
-        "best_optimized.pt",
-    ),
-)
+    # Saving model
+    os.makedirs(OUTPUT_DIR, exist_ok=True)
+    ckpt = {
+        "model": model.model,
+        "train_args": {},
+    }
+    torch.save(
+        ckpt,
+        os.path.join(
+            OUTPUT_DIR,
+            "best_optimized.pt",
+        ),
+    )
+    print("Pruning completed")
+except Exception as e:
+    exc_type, exc_value, exc_traceback = sys.exc_info()
+    tb = traceback.extract_tb(exc_traceback)
+    filename, line_number, func_name, text = tb[-1]
+    print(f"Error occurred in file: {filename}")
+    print(f"Line {line_number}: {text}")
+    print(f"Error message: {e}")
